task 104

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,19 +9,19 @@
 int main(void)
 {
 	int i;
-float fibonacci[98];
+long double fibonacci[98];
 
 	fibonacci[0] = 1;
 	fibonacci[1] = 2;
-	printf("%f, %f, ", fibonacci[0], fibonacci[1]);
+	printf("%ld, %ld, ", fibonacci[0], fibonacci[1]);
 
 	for (i = 2; i < 98; i++)
 	{
 		fibonacci[i] = fibonacci[i - 1] + fibonacci[i - 2];
 		if (i == 97)
-			printf("%f\n", fibonacci[i]);
+			printf("%ld\n", fibonacci[i]);
 		else
-			printf("%f, ", fibonacci[i]);
+			printf("%ld, ", fibonacci[i]);
 	}
 
 	return (0);
